+++ src/lib.rs
@@ -12,10 +12,12 @@ type Result<T> = std::result::Result<T, Box<dyn error::Error>>;
 
 #[cfg(test)]
 mod tests{
-    use crate::EncryptoRSA;
+    use num_bigint::BigUint;
+    use crate::{EncryptoRSA, Generator, One};
 
     #[test]
     fn encrypto_tests(){
+        //29329910984667111668698655637818737335150278953697889411613371939678947564721136189355438568499715776963467791924170712468382877544862676115129685014796720016362374716002936880408902090892990110715659120207495854365051339106405607318567460641209540060903987224813697125337686240065295060748518483892816811089217073863560993386552806111000416661028596519930847932759070529862887234415643686990441277249722114110301213080169821499693026604664073773056594900899391540646777145266698605869343033207736619528293708523985002390293641933265245787779900591773848698907339399218748495990024797759753087538276697216466593947233
         let encrypto = EncryptoRSA::init(512);
         let encrypto1 = EncryptoRSA::init(512);
         let msg = b"abc".as_slice();
@@ -23,13 +25,60 @@ mod tests{
         let enc = encrypto.double_encrypt(msg, EncryptoRSA::desterilize_pub_key(encrypto1.get_sterilized_pub_key())).unwrap();
         let dec = encrypto1.double_decrypt(enc.as_bytes(), encrypto.pbl.clone());
     }
+
+    #[test]
+    fn cross_platform(){
+        let encrypto = EncryptoRSA::init(512);
+        println!("{}", encrypto.get_sterilized_pub_key());
+        println!("Enter message: ");
+        let mut msg = String::new();
+        std::io::stdin().read_line(&mut msg).unwrap();
+        println!("Choose decryption method decrypt, double_decrypt, double_decrypt_with_pkcsv1_15, decrypt_with_pkcsv1_15: ");
+        let mut dec_method = String::new();
+        std::io::stdin().read_line(&mut dec_method).unwrap();
+
+        if "decrypt".to_string() == dec_method {
+            println!("{:?}",encrypto.decrypt(msg));
+        }else if "double_decrypt".to_string() == dec_method {
+            let mut pubkey = String::new();
+            std::io::stdin().read_line(&mut pubkey).unwrap();
+            println!("{:?}",encrypto.double_decrypt(msg.as_bytes(), EncryptoRSA::desterilize_pub_key(pubkey)));
+        }else if "double_decrypt_with_pkcsv1_15".to_string() == dec_method {
+            let mut pubkey = String::new();
+            std::io::stdin().read_line(&mut pubkey).unwrap();
+            println!("{:?}",encrypto.double_decrypt_with_pkcsv1_15(msg, EncryptoRSA::desterilize_pub_key(pubkey)));
+        }else if "decrypt_with_pkcsv1_15".to_string() == dec_method {
+            println!("{:?}", encrypto.decrypt_with_pkcsv1_15(msg));
+        }
+    }
+
 }
 
 #[cfg(test)]
 mod private_tests{
-    use num_bigint::{BigUint, ToBigInt};
-    use crate::{Generator, modinv, One};
+    use num_bigint::{BigUint, RandBigInt, ToBigInt};
+    use crate::{EncryptoRSA, Generator, modinv, One};
 
+    #[test]
+    fn b64check(){
+        //[67, 84, 116, 100, 77, 88, 87, 73, 100, 72, 103, 73, 84, 85, 119, 83, 72, 50, 74, 51, 67, 122, 50, 76, 56, 111, 85, 104, 49, 83, 82, 104, 120, 112, 103, 79, 72, 99, 80, 73, 73, 50, 69, 61, 10]
+        let b64 = base64::decode("eyJwZSI6NjU1MzcsIm4iOjcwOTQ0NzU1MDgwNjM0OTIzNzI3NTQ0NzkwMjM0Mzg3MzE0NTQ4NTQ0OTIzNjAwMjQ1NzA1NjM1ODYxNDg4NjIxNDU3MzgwOTI3MjA3LCJrZXlsZW4iOjEyOH0=").unwrap();
+        // let b641 = base64::decode("aLTdPaIP8v5YA3bwETeM4OP88/+28dfzkshLkP5tOqk=".as_bytes()).unwrap();
+        //
+        let v = [101, 121, 74, 119, 90, 83, 73, 54, 78, 106, 85, 49, 77, 122, 99, 115, 73, 109, 52, 105, 79, 106, 99, 119, 79, 84, 81, 48, 78, 122, 85, 49, 77, 68, 103, 119, 78, 106, 77, 48, 79, 84, 73, 122, 78, 122, 73, 51, 78, 84, 81, 48, 78, 122, 107, 119, 77, 106, 77, 48, 77, 122, 103, 51, 77, 122, 69, 48, 78, 84, 81, 52, 78, 84, 81, 48, 79, 84, 73, 122, 78, 106, 65, 119, 77, 106, 81, 49, 78, 122, 65, 49, 78, 106, 77, 49, 79, 68, 89, 120, 78, 68, 103, 52, 78, 106, 73, 120, 78, 68, 85, 51, 77, 122, 103, 119, 79, 84, 73, 51, 77, 106, 65, 51, 76, 67, 74, 114, 90, 88, 108, 115, 90, 87, 52, 105, 79, 106, 69, 121, 79, 72, 48, 61];
+        assert_eq!("eyJwZSI6NjU1MzcsIm4iOjcwOTQ0NzU1MDgwNjM0OTIzNzI3NTQ0NzkwMjM0Mzg3MzE0NTQ4NTQ0OTIzNjAwMjQ1NzA1NjM1ODYxNDg4NjIxNDU3MzgwOTI3MjA3LCJrZXlsZW4iOjEyOH0=".as_bytes(), v);
+        let xx = "eyJrZXlsZW4iOiIxMjgiLCJuIjoiNzI3ODQwMjEwOTAyMzcxMzMyNjQ1NzMzNDQ2MTQzMTk3NTQxODE0MTEwNzMyNzI1NzQ4MjM1MjY4NjY5OTM3OTI4ODI0OTUwODE1MSIsInBlIjoiNjU1MzcifQ==";
+        let pk = EncryptoRSA::desterilize_pub_key(xx.to_string());
+        println!("{:?}", pk);
+        let x = "eyJwZSI6NjU1MzcsIm4iOjcwOTQ0NzU1MDgwNjM0OTIzNzI3NTQ0NzkwMjM0Mzg3MzE0NTQ4NTQ0OTIzNjAwMjQ1NzA1NjM1ODYxNDg4NjIxNDU3MzgwOTI3MjA3LCJrZXlsZW4iOjEyOH0=".to_string();
+        let pk = EncryptoRSA::desterilize_pub_key(x);
+    }
+
+    #[test]
+    fn check_len(){
+        let mut v = rand::thread_rng().gen_biguint(128).to_bytes_le();
+        println!("{:?}", v);
+    }
     #[test]
     fn idk(){
         let bit_len = 1024  ;
@@ -40,7 +89,6 @@ mod private_tests{
         let n = p.clone() * q.clone();
         let on = (p - BigUint::one()) * (q - BigUint::one());
         let d = modinv(e.clone().to_bigint().unwrap(), on.clone().to_bigint().unwrap()).unwrap();
-        assert_eq!(BigUint::one(), (d.clone()*e.clone())%on.clone());
 
         //alice
         let p1 = Generator::new_prime(bit_len);
@@ -114,8 +162,9 @@ impl EncryptoRSA {
         let on = (p - BigUint::one()) * (q - BigUint::one());
         let d = modinv(e.clone().to_bigint().unwrap(), on.clone().to_bigint().unwrap()).unwrap();
 
-        assert_eq!(BigUint::one(), (d.clone() * e.clone()) % on);
-
+        if BigUint::one() != (d.clone()*e.clone())%on.clone() {
+            return EncryptoRSA::init(bit_len);
+        }
         let pbl: ZotPublicKey = ZotPublicKey {
             e,
             n: n.clone(),
@@ -147,13 +196,11 @@ impl EncryptoRSA {
     pub fn desterilize_pub_key(encoded: String) -> ZotPublicKey {
         let x = base64::decode(encoded).unwrap();
         let json: Value = serde_json::from_slice(&*x).unwrap();
-        let x = json.get("n").unwrap().as_str().unwrap();
-        let xx = x.as_bytes();
-        let n = BigUint::parse_bytes(xx, 10).unwrap();
-        let x = json.get("pe").unwrap().as_str().unwrap();
-        let xx = x.as_bytes();
-        let e = BigUint::parse_bytes(xx, 10).unwrap();
-        let bit_len = usize::from_str(json.get("pe").unwrap().as_str().unwrap()).unwrap();
+        let x = json.get("n").unwrap().as_str().unwrap().as_bytes();
+        let n = BigUint::parse_bytes(x, 10).unwrap();
+        let x = json.get("pe").unwrap().as_str().unwrap().as_bytes();
+        let e = BigUint::parse_bytes(x, 10).unwrap();
+        let bit_len = usize::from_str(json.get("keylen").unwrap().as_str().unwrap()).unwrap();
         ZotPublicKey {
             e,
             n,
@@ -166,30 +213,28 @@ impl EncryptoRSA {
         let mut hm = HashMap::<&str, String>::new();
         hm.insert("pe", self.pbl.e.clone().to_string());
         hm.insert("n", self.pbl.n.clone().to_string());
+        hm.insert("keylen", self.pbl.keylen.clone().to_string());
         let json = serde_json::to_value(hm).unwrap().to_string();
-        base64::encode(json.as_bytes())
+        base64::encode(json.to_string().as_bytes())
     }
 
     /// This method adds random bytes to the message, encrypts with the `pub_key` and again encrypts it with your private key.
     ///
     /// You can decrypt it using double_decrypt(...) method
-    pub fn double_encrypt(&self, bytes: &[u8], pub_key: ZotPublicKey) -> Result<String> {
+    pub fn double_encrypt<T: AsRef<[u8]>>(&self, bytes: T, pub_key: ZotPublicKey) -> Result<String> {
         if pub_key.keylen - 11 < bytes.len() {
             panic!("Msg bigger than key-length, use at least 2048 bit key");
         }
         let bi = convert_bytes_to_big_int(bytes);
         let enc = bi.modpow(&pub_key.e, &pub_key.n);
-        println!("{}", enc.clone());
         let enc = enc.modpow(&self.pri.d, &self.pri.n);
-        println!("{}", enc.clone());
-        println!("{}", bi);
         Ok(base64::encode(convert_bigint_to_bytes(enc)))
     }
 
     /// This method adds random bytes to the message, encrypts with the `pub_key`.
     ///
     /// You can decrypt it using decrypt_with_pkcsv1_15(...) method
-    pub fn encrypt_with_pkcsv1_15(&self, bytes: &[u8], pub_key: ZotPublicKey) ->  Result<String> {
+    pub fn encrypt_with_pkcsv1_15<T: AsRef<[u8]>>(&self, bytes: T, pub_key: ZotPublicKey) ->  Result<String> {
         if pub_key.keylen - 11 < bytes.len() {
             panic!("Msg bigger than key-length, use at least 2048 bit key");
         }
@@ -203,7 +248,7 @@ impl EncryptoRSA {
     /// This method adds random bytes to the message, encrypts with the `pub_key` and again encrypts it with your private key.
     ///
     /// You can decrypt it using double_decrypt_with_pkcsv1_15(...) method
-    pub fn double_encrypt_with_pkcsv1_15(&self, bytes: &[u8], pub_key: ZotPublicKey) -> Result<String> {
+    pub fn double_encrypt_with_pkcsv1_15<T: AsRef<[u8]>>(&self, bytes: T, pub_key: ZotPublicKey) -> Result<String> {
         if pub_key.keylen - 11 < bytes.len() {
             panic!("Msg bigger than key-length, use at least 2048 bit key");
         }
@@ -218,7 +263,7 @@ impl EncryptoRSA {
     ///This method encrypts with the `pub_key`.
     ///
     /// You can decrypt it using decrypt(...) method
-    pub fn encrypt(&self, bytes: &[u8], pub_key: ZotPublicKey) ->  Result<String> {
+    pub fn encrypt<T: AsRef<[u8]>>(&self, bytes: T, pub_key: ZotPublicKey) ->  Result<String> {
         if pub_key.keylen - 11 < bytes.len() {
             panic!("Msg bigger than key-length, use at least 2048 bit key");
         }
@@ -230,7 +275,7 @@ impl EncryptoRSA {
     ///This method decrypts value twice, once with public key and then with private key.
     ///
     /// this way you know that the public key is from the designated sender
-    pub fn double_decrypt(&self, val: &[u8], pub_key: ZotPublicKey) -> Vec<u8> {
+    pub fn double_decrypt<T: AsRef<[u8]>>(&self, val: T, pub_key: ZotPublicKey) -> Vec<u8> {
         let by = base64::decode(val).unwrap();
         let bi = convert_bytes_to_big_int(&*by);
         let dec = bi.modpow(&pub_key.e, &pub_key.n);
@@ -243,7 +288,7 @@ impl EncryptoRSA {
     /// this way you know that the public key is from the designated sender
     ///
     /// It removes first 16 bytes, which were used to increase message length to protect it against attacks
-    pub fn double_decrypt_with_pkcsv1_15(&self, val: String, pub_key: ZotPublicKey) -> Vec<u8> {
+    pub fn double_decrypt_with_pkcsv1_15<T: AsRef<[u8]>>(&self, val: T, pub_key: ZotPublicKey) -> Vec<u8> {
         let by = base64::decode(val).unwrap();
         let bi = convert_bytes_to_big_int(&*by);
         let dec = bi.modpow(&pub_key.e, &pub_key.n);
@@ -254,21 +299,23 @@ impl EncryptoRSA {
     }
 
     /// This method decrypts value with private key.
-    pub fn decrypt(&self, val: &[u8]) -> Vec<u8> {
-        let by = base64::decode(val).unwrap();
-        let bi = convert_bytes_to_big_int(&*by);
+    pub fn decrypt<T: AsRef<[u8]>>(&self, val: T) -> Vec<u8> {
+        // let val = val.as_bytes();
+        let by = &base64::decode(val).unwrap()[..];
+        // println!("{:?} {:?}", val, by.clone());
+        let bi = convert_bytes_to_big_int(by);
         // let dec = (bi*self.pbl.e.clone()) % self.pbl.n.clone();
         let dec = bi.modpow(&self.pri.d.clone(), &self.pri.n.clone());
         convert_bigint_to_bytes(dec)
     }
 
     /// This method decrypts value with private key and returns decoded value of pkcsv1 1.5 padding.
-    pub fn decrypt_with_pkcsv1_15(&self, val: &[u8]) -> Vec<u8> {
-        let by = base64::decode(val).unwrap();
+    pub fn decrypt_with_pkcsv1_15<T: AsRef<[u8]>>(&self, val: T) -> Vec<u8> {
+        /*let by = base64::decode(val).unwrap();
         let bi = convert_bytes_to_big_int(&*by);
         // let dec = (bi*self.pbl.e.clone()) % self.pbl.n.clone();
-        let dec = bi.modpow(&self.pri.d.clone(), &self.pri.n.clone());
-        let mut x = convert_bigint_to_bytes(dec);
+        let dec = bi.modpow(&self.pri.d.clone(), &self.pri.n.clone());*/
+        let mut x = self.decrypt(val);
         x.drain(0..16);
         x
     }